# cmd/gameserverapi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service for managing GameServer and GameServerBuilds",
        "title": "GameServer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/gameserverbuilds": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get list of GameServerBuilds",
                "operationId": "get-list-gameserverbuilds",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1alpha1.GameServerBuildList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Create a GameServerBuild",
                "operationId": "create-gameserverbuild",
                "parameters": [
                    {
                        "description": "gsb",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1alpha1.GameServerBuild"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v1alpha1.GameServerBuild"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/gameserverbuilds/{namespace}/{buildName}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get a GameServerBuild by buildName and namespace",
                "operationId": "get-game-server-build-by-buildName\u0026namespace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespaceParam",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "buildNameParam",
                        "name": "buildName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1alpha1.GameServerBuild"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "delete GameServerBuild by buildName and namespace",
                "operationId": "path-gameserverbuild-by-buildname-and-namespace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "buildNameParam",
                        "name": "buildName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespaceParam",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "summary": "patch GameServerBuild by buildName and namespace",
                "operationId": "path-gameserverbuild-by-buildname-and-namespace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "buildNameParam",
                        "name": "buildName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespaceParam",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "gsbMap",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1alpha1.GameServerBuild"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/gameserverbuilds/{namespace}/{buildName}/gameserverdetails": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get GameServerDetailList by buildName",
                "operationId": "get-gameserver-details-by-build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "buildNameParam",
                        "name": "buildName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespaceParam",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1alpha1.GameServerDetailList"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/gameserverbuilds/{namespace}/{buildName}/gameservers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get list of GameServers for a given build",
                "operationId": "get-list-gameservers-by-build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "buildNameParam",
                        "name": "buildName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespaceParam",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1alpha1.GameServerList"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/gameserverbuilds/{namespace}/{gameServerDetailName}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get GameServerDetail by GameServerDetailName and namespace",
                "operationId": "get-gameserver-details-by-gameserverdetailname-and-namespace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gameServerDetailNameParam",
                        "name": "gameServerDetailName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespaceParam",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1alpha1.GameServerDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/gameservers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get list of GameServers",
                "operationId": "get-list-gameservers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1alpha1.GameServerList"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/gameservers/{namespace}/{gameServerName}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get GameServer by GameServerName and namespace",
                "operationId": "get-gameserver-by-gameservername-and-namespace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gameServerNameParam",
                        "name": "gameServerName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespaceParam",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1alpha1.GameServer"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "delete GameServer by GameServerName and namespace",
                "operationId": "delete-gameserver-by-gameservername-and-namespace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gameServerNameParam",
                        "name": "gameServerName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "namespaceParam",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
...
        "v1alpha1.GameServerBuildList": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "description": "APIVersion defines the versioned schema of this representation of an object.\nServers should convert recognized schemas to the latest internal value, and\nmay reject unrecognized values.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources\n+optional",
                    "type": "string"
                },
                "continue": {
                    "description": "continue may be set if the user set a limit on the number of items returned, and indicates that\nthe server has more data available. The value is opaque and may be used to issue another request\nto the endpoint that served this list to retrieve the next set of available objects. Continuing a\nconsistent list may not be possible if the server configuration has changed or more than a few\nminutes have passed. The resourceVersion field returned when using this continue value will be\nidentical to the value in the first response, unless you have received this token from an error\nmessage.",
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1alpha1.GameServerBuild"
                    }
                },
                "kind": {
                    "description": "Kind is a string value representing the REST resource this object represents.\nServers may infer this from the endpoint the client submits requests to.\nCannot be updated.\nIn CamelCase.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds\n+optional",
                    "type": "string"
                },
                "remainingItemCount": {
                    "description": "remainingItemCount is the number of subsequent items in the list which are not included in this\nlist response. If the list request contained label or field selectors, then the number of\nremaining items is unknown and the field will be left unset and omitted during serialization.\nIf the list is complete (either because it is not chunking or because this is the last chunk),\nthen there are no more remaining items and this field will be left unset and omitted during\nserialization.\nServers older than v1.15 do not set this field.\nThe intended use of the remainingItemCount is *estimating* the size of a collection. Clients\nshould not rely on the remainingItemCount to be set or to be exact.\n+optional",
                    "type": "integer"
                },
                "resourceVersion": {
                    "description": "String that identifies the server's internal version of this object that\ncan be used by clients to determine when objects have changed.\nValue must be treated as opaque by clients and passed unmodified back to the server.\nPopulated by the system.\nRead-only.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency\n+optional",
                    "type": "string"
                },
                "selfLink": {
                    "description": "Deprecated: selfLink is a legacy read-only field that is no longer populated by the system.\n+optional",
                    "type": "string"
                }
            }
        },
...
    }
}